requirements.txt
asyncwhois
colorama
colored
coloredlogs
dnspython
folium
geoip2
ruamel.yaml
simplejson
tqdm
verboselogs